✨ Refactor Playwright tests structure and enhance drush command handling
