+++ package.json
@@ -4,11 +4,13 @@
   "description": "Playwright end-to-end tests for prisoner-content-hub-backend",
   "scripts": {
     "test:e2e": "playwright test",
+    "test:e2e:all": "playwright test --workers=1",
     "test:e2e:headed": "playwright test --headed",
     "test:e2e:debug": "playwright test --debug",
     "test:e2e:ui": "playwright test --ui",
     "test:e2e:report": "playwright show-report",
-    "typecheck:e2e": "tsc --noEmit -p tsconfig.playwright.json"
+    "typecheck:e2e": "tsc --noEmit -p tsconfig.playwright.json",
+    "test:all": "npm run typecheck:e2e && npm run test:e2e:all"
   },
   "devDependencies": {
     "@playwright/test": "^1.55.0",
